Add reference started game view
